+++ src/Symfony/Component/Serializer/Tests/Fixtures/EmbeddedDummyChild.php
@@ -1,0 +1,28 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Serializer\Tests\Fixtures;
+
+/**
+ * @author Jibé Barth <barth.jib@gmail.com>
+ */
+class EmbeddedDummyChild
+{
+    /**
+     * @var string
+     */
+    public $bar;
+
+    /**
+     * @var string
+     */
+    public $baz;
+}